refactor: move budgeting::user_budget types from lib to wire
