+++ lib/src/budgeting/user_budget.rs
@@ -1,29 +1,16 @@
-use crate::common::{is_false, request, request_bare, SerializableNone};
+use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
 use anyhow::Context;
 use chrono::{DateTime, FixedOffset};
+use lrzcc_wire::budgeting::{
+    UserBudget, UserBudgetCombined, UserBudgetCombinedDetail,
+    UserBudgetCreateData, UserBudgetDetail, UserBudgetModifyData,
+    UserBudgetOver,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserBudget {
-    pub id: u32,
-    pub user: u32,
-    pub username: String,
-    pub year: u32,
-    pub amount: u32,
-}
-
-impl Display for UserBudget {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("UserBudget(id={})", self.id))
-    }
-}
 
 pub struct UserBudgetApi {
     pub url: String,
@@ -102,25 +89,6 @@ impl UserBudgetListRequest {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-struct UserBudgetCreateData {
-    user: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    year: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    amount: Option<i64>,
-}
-
-impl UserBudgetCreateData {
-    fn new(user: u32) -> Self {
-        Self {
-            user,
-            year: None,
-            amount: None,
-        }
-    }
-}
-
 pub struct UserBudgetCreateRequest {
     url: String,
     client: Rc<Client>,
@@ -155,26 +123,6 @@ impl UserBudgetCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct UserBudgetModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    amount: Option<u32>,
-    #[serde(skip_serializing_if = "is_false")]
-    force: bool,
-}
-
-impl UserBudgetModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            amount: None,
-            force: false,
-        }
     }
 }
 
@@ -213,50 +161,6 @@ impl UserBudgetModifyRequest {
             StatusCode::OK,
         )
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserBudgetOver {
-    pub budget_id: u32,
-    pub user_id: u32,
-    pub user_name: String,
-    pub over: bool,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserBudgetCombined {
-    pub budget_id: u32,
-    pub user_id: u32,
-    pub user_name: String,
-    pub project_budget_id: u32,
-    pub project_id: u32,
-    pub project_name: String,
-    pub over: bool,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserBudgetDetail {
-    pub budget_id: u32,
-    pub user_id: u32,
-    pub user_name: String,
-    pub over: bool,
-    pub cost: f64,
-    pub budget: u32,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserBudgetCombinedDetail {
-    pub budget_id: u32,
-    pub user_id: u32,
-    pub user_name: String,
-    pub project_budget_id: u32,
-    pub project_id: u32,
-    pub project_name: String,
-    pub over: bool,
-    pub project_cost: f64,
-    pub project_budget: u32,
-    pub user_cost: f64,
-    pub user_budget: u32,
 }
 
 #[derive(Debug)]
